LU-9313 ldlm: Limit elc to one thread per namespace
Doing ELC from many threads at once can causes soft lockups
due to contention on the ns_lock as all the threads attempt
to build lists of locks to cancel.  (ldlm_prepare_lru_list)

This patch limits ELC to one thread at a time per
namespace.  This prevents the soft lockups without harming
lock cancellation much.  ELC is an opportunistic mechanism,
and if it is unable to keep up with lock creation rates,
the ldlm_bl threads will handle lock cancellation if
necessary.

Upstream:
https://review.whamcloud.com/#/c/26477/2/

Note we are taking patch set 2, because Vitaly supports the
idea, but asked for this to be moved around and rebased on
other changes which haven't landed yet, but we need a fix
now.

We also remove the ldlm_cli_cancel_list call in mdc_rename.
This call is incorrect, and is usually a no-op because
ldlm_prep_elc_req empties the "cancels" list.

When we don't do that, this call happens, and it panics.
This is the only place in the code where such a call is
made after doing elc prep.

Bugs Affected: 849982
Bug Action: UPDATE