LU-6952 mount: Mount options parsing problem
The code which parse the mount option expect that the default
mount options come first and then the user given options. But
actually the default options are appended after the user given
option. Because of this user given options are cleared.
Following code fix this bug.

Added a conf-sanity test 104 to verify this fix.

This patch is back-ported from the following one:
Lustre-commit: b27652cb440f2bea59db254e955403484c6c1d3a
Lustre-change: http://review.whamcloud.com/#/c/15906/

Change-Id: I1a22e5d559e01d8bcba0ceebba5544e92d39bb19
Seagate-bug-id: MRP-2193
Signed-off-by: Pratik Shinde <pratik.shinde@seagate.com>
Signed-off-by: Rahul Deshmukh <rahul.deshmukh@seagate.com>
Signed-off-by: Hongchao Zhang <hongchao.zhang@intel.com>
Change-Id: I77418b071413b80c8c246306128ab0c348b41fa8
Reviewed-on: https://review.whamcloud.com/27801
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jian Yu <jian.yu@intel.com>
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Tested-by: Jenkins