848035 build: Add spec file for CLE VM
The CLE VM project is intended to create a CLE-like environment but running on
VMs.  The build environment is defined at
https://connect.us.cray.com/confluence/display/CLEVM/Build+Environment.  For
the current iteration of CLE VM, the kernel flavor is default.  This mod adds
in a spec file to the lustre source to enable lustre builds for the CLE VM
environment.  It is a first pass attempt in order to establish a basic build
environment.

Bugs Affected: 848035
Bug Action: RESOLVE