851037 ptlrpc: Use 1 MiB RPCs by default
Commit "062a27d907 LU-7990 clio: revise readahead to support 16MB IO"
changed the default RPC size to 16 MB (negotiable to the maximum
supported by the server). This had the side effect of exposing some
bugs on an internal test filesystem as documented in bug 850696.
While we would like to change the default RPC size for the improved
performance, we've decided not to do so for the current release in
development (Lustre 2.7.2).

Bugs Affected: 851037
Bug Action: RESOLVE