851037 tests: Set 16MB RPC size via lctl for 2.7.2
The sanity.sh test 101g assumes that the client will attempt to use
the largest RPC size supported by the server, so it sets 16MB RPC
size on the server and remounts the client so the client picks up the
larger size. Since we've changed that behavior for 2.7.2 we need to
explicitly set 16MB RPC size via lctl on the client.

Bugs Affected: 851037
Bug Action: UPDATE