Make changes to lustre metaspec builds for dvs
LUS-68, LUS-57, LUS-113

Dvs makes assumptions about where certain things are installed such as
the Modules.symvers file. I changed the default prefix to /usr because
hat is where the upstream lustre builds install things. The
versioning was also causing installation problems so I changed it to
something a little more sane. The new packages look like this:

lustre-client-compute-2.7.3.35-1.0000.8f363db.0.0.x86_64.rpm

Which conforms with the BITS group versioning scheme. The versioning
is also now consistant across all packages produced by the build which
was previously not the case. I also removed ownership of the root
file system because that was creating installation conflicts.

I put a copy of the Module.symvers file in the cray-lustre-...-module
package. This allows for the installation of multipe Module.symvers
files in the obs build root for dvs. Dvs requires the Module.symvers
file from the ari_s and ari_c builds. If left in the default locatin
it is impossilbe to install multiple Module.symvers files on the same
system because they conflict.