gnilnd: Handle seq_printf api change
Starting with the 4.3 linux kernel seq_printf is a void
function not returning the character count. Upstream Lustre
did not update all of gnilnd to accomodate this change
when they updated the rest of Lustre. This change takes
care of what was missed.

This commit can be removed from Cray-master via rebase when the
upstream source is fixed.