LUS-150 LUS-151 nullify badness rating
The sles12sp2 and sles12sp3 builds are currently failing because the
badness rating is too high. This is because the build does not want
spec files in the contents of rpms. Specifically the
lustre...source...rpm contains the spec file and the spec.in file
which is redundant when you take into account the src.rpm. I am
removing the lustre.spec and lustre.spec.in files from the source
package