849628 gnilnd: Use consistent size for GNI MBOX.
KMALLOC_MAX_SIZE is not consistent when changing
the allocator. For example when using the SLUB allocator
the size is set to 30 bits. For slab allocator the max size
is 25 bits or 32MB.

For SLAB this is a no-op change for other allocators the size
is also set to 32MB.

Bugs Affected: 849628
Bug Action: RESOLVE