LU-8014 hsm: remove invalid kuc_free in coordinator
Commit 9091a934fc ("LU-8014 hsm: remove kuc_ispayload") moved the KUC
allocation in mdt_hsm_agent_send() and I left a kuc_free in an error
path in mdt_coordinator(). The HAL is no longer a KUC and it is freed
outside the for loop. That kuc_free is incorrect.

This patch is back-ported from the following one:
Lustre-commit: 12ac7f2baa3691eaf74d468aedb7615f6ddf6d60
Lustre-change: https://review.whamcloud.com/19869

Signed-off-by: frank zago <fzago@cray.com>
Change-Id: If0894d1a8bcbc7960b728e33a34bddb95806acf8
Reviewed-on: https://review.whamcloud.com/28349
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Faccini Bruno <bruno.faccini@intel.com>
Reviewed-by: John L. Hammond <john.hammond@intel.com>