CAST-7745: eliminate modifying rpcs from path2fid
Robinhood scans spend too much time executing path2fid. The bottleneck
in path2fid is the limit on the number of modifying rpcs in flight to
the
MDS. If the server does not support multiple modifying rpcs (pre-2.7.2)
then the path2fid operations are serialized at several points - file
open,
close, and getxattr.

The scan should not be modifying the file system (with perhaps the
exception of access time) so its Lustre accesses should not need to be
tagged as modifying nor reserve one of the limited modifying rpcs in
flight.

This mod removes read-only OPENs, their matching CLOSEs, and GETXATTR
from the Lustre operations that are considered modifying and that
require
a slot in the table of modifying rpcs in flight. It also adds scan_only
to
/proc/sys/lustre so this scan only behavior can be dynamically enabled/
disabled. The default is disabled, which results in the behavior prior
to
this mod.