LUS-102 build: Implement new packaging
This commit implements a new spec file for OBS builds.
The cray-lustre.spec file:
    - Generates the following packages depending on build project:
        - cray-lustre-kmp-cray_ari_s.rpm
            - Kernel modules for aries service nodes
            - Includes server modules
        - cray-lustre-kmp-cray_ari_c.rpm
            - Kernel modules for aries compute nodes
        - cray-lustre-kmp-default.rpm
            - Kernel modules for elogin, or sles whitebox server
        - cray-lustre-{service,compute,elogin,server}.rpm
            - /etc, /bin, /sbin, etc.
        - cray-lustre-{service,compute,elogin,server}-devel.rpm
            - *.a
            - *.so
            - liblustreapi.h, etc.
            - *.pc files for building against liblustreapi
            - module file (service and default flavor builds only)
        - cray-lustre-{service,compute}-lnet-devel.rpm
            - headers for building against lnet (service only)
            - cray-lnet.pc file for building against LNet (service only)
            - Module.symvers

On XC Compute nodes the following packages should be installed:
        - cray-lustre-kmp-cray_ari_c.rpm
        - cray-lustre-compute.rpm
        - cray-lustre-compute-devel.rpm

On XC Service nodes the following packages should be installed:
        - cray-lustre-kmp-cray_ari_s.rpm
        - cray-lustre-service.rpm
        - cray-lustre-service-devel.rpm

On eLogin nodes the following packages should be installed:
        - cray-lustre-kmp-default.rpm
        - cray-lustre-elogin.rpm
        - cray-lustre-elogin-devel.rpm

The cray-lustre-{service,compute}-lnet-devel.rpm packages will only be
used to build DVS in OBS.

New pkg-config files referenced by the new spec files are also added
in this commit.

Lustre already provides a systemd unit file for lnet, but we need to
add it to our package, and modify it to work correctly in Cray
environment.

This spec file should also support the CLE VM environment.

Files are installed into standard locations under /usr. The easiest way
to do this is to %define _prefix to /usr in the spec file. OBS macros
will then define the other directory locations, e.g. _libdir, under
/usr.

A note about the kmp package names:
As documented in LUS-276 there's a desire to be able to determine the
obs build command used to generate kmp packages from the information
provided by the package itself. The rpm-info 'distribution' entry
tells us the obs build project and obs build repo, but we also need
the spec file used for the build. In the case of other Cray kernel
module packages the rpm-info 'name' field is like:
cray-lustre-kmp-cray_ari_c
This allows scripts to recognize the '-kmp-' substring and determine
the build spec is the substring preceding '-kmp-'. e.g.
cray-lustre.spec

Since we want our base package and subpackages to be distinguishable
by the platform they are built for we don't want to change the base
package name. Thus, we override the kmp package name using the '-n'
option to the appropriate %kernel_module_package macro.

Also note that we utilize the '-v' option to the aforementioned
macro. This is to avoid having the kernel version appear twice in the
kmp package version which will happen by default because we include
the kernel version string in the base package version.