LUS-1718 build: Correctly detect clevm build
The CRAY:Network:VM build project is linked into the SLE_12*_VM
repositories to tell us that we're doing a clevm build. However, this
build project just has '%define clevm 1' in it's project config rather
than '%with_clevm 1', so we need to tweak the logic used to detect
whether we're doing a clevm build in the cray-lustre.spec file.