LUS-1737 build: Modify files section in ari_s-rhine.spec
The debug builds fail because the current files section in the
spec file results in debug artifacts being packaged in the base
rpm rather than the debuginfo rpm.

Goal of this mod is to resolve this issue without making any additional
packaging changes. i.e. the list of files in the packages created by
this spec file should be identical before and after this mod.