LUS-5911 libcfs: remove mem wrappers for libcfs
This is a partial port of upstream commit
6ea4de1b7f237d331c9e0d66f4cc53365d036e5d
(https://review.whamcloud.com/13841)

The libcfs layer contains wrappers to handle memory
management for both kernel space and user land. Since
libcfs is no longer built for user land we can remove
the wrappers. For linux-mem.h we kept only what is
needed to deal with different kernel versions. The
rest was moved to libcfs_prim.h.

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: Ie198500cdfa28f5d4dc8635ce0a41726ab158cbb
Intel-bug-id: LU-6245
Reviewed-on: http://review.whamcloud.com/13841
Reviewed-by: frank zago <fzago@cray.com>
Tested-by: Jenkins
Reviewed-by: Dmitry Eremin <dmitry.eremin@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: John L. Hammond <john.hammond@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>
Reviewed-on: https://es-gerrit.dev.cray.com/17798
Reviewed-by: Alexander Boyko <c17825@cray.com>
Reviewed-by: Andriy Skulysh <c17819@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>