LUS-149 build: Symlink removed on rpm upgrade
ko2iblnd-probe and lctl symlinks are unconditionally created in %post
and removed in %preun. This causes the symlinks to be removed on rpm
upgrade because the new package gets installed first before the old
one is removed. Instead create the links in the %install section and
package them like any other file. This way they are actually owned by
the package.

Signed-off-by: Chris Horn <hornc@cray.com>