LUS-6252 obdecho: don't mask ma_need in echo_attr_get_complex()
echo_attr_get_complex() copies ma_need to a local variable, masks
MA_* values other than MA_INODE if MA_INODE is set in ma_need,
and restores the saved value of ma_need before the function exits.
This does not seem to be useful, and triggers an assertion in
echo_big_lmm_get() when MA_LOV and/or MA_LMV is set in ma_need.

Change-Id: Ia33c4cd2c9f9cff2045ac99fb4f3491efdb1ca9e
Signed-off-by: Nikitas Angelinas <nangelinas@cray.com>
Reviewed-on: https://es-gerrit.dev.cray.com/153636
Reviewed-by: Patrick Farrell <paf@cray.com>
Reviewed-by: Andrew Perepechko <c17827@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>