LUS-6455 libcfs: crashes with certain cpu part numbers
Due to a bug in the code, libcfs will crash if the
number of online cpus does not divide by the number
of cpu partitions.

Based on the checks in cfs_cpt_table_create(), it
appears that the original intent was to push the
remaining cpus into the initial partitions.

So let's do that properly and put a console warning
in this case because such a configuration is likely
suboptimal.

Change-Id: I5c1a05cc8475d1eb9660091206b2bcd18f0df3f0
Signed-off-by: Andrew Perepechko <c17827@cray.com>
Reviewed-on: https://es-gerrit.dev.cray.com/153963
Reviewed-by: Alexander Boyko <c17825@cray.com>
Reviewed-by: Alexander Zarochentsev <c17826@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>