LUS-6565 kfilnd: Initial Prototype from cassini-lnd
This is a direct copy of the prototype in the cassini-lnd project.
All patches that went into this prototype have been inspected and
run on an emulator.  The prototype is being pushed into the Lustre
repo where work will continue on it.

Note: this first patch is a direct copy and does not build in the
Lustre repo.  A following patch will introduce the changes required
for building and running this prototype from the Lustre repo.

Signed-off-by: Doug Oucharek <doucharek@cray.com>
Change-Id: Ie92f3d6b8b062365b25064115def3f97598b05d6