LUS-6843 hsm: Create new copytool interface
The current copytool interface has a combination of opaque
and visible data, which increases, rather than hides
complexity. This seeks to make a parallel interface
for start, end and progress which is simpler.

The new interface uses the files that are being
operated on as the targets for the IOCTL, rather
than the root directory.

lhsmtool_posix: New mode which takes an input file of work.
This new file is specified with the -F option.  Each line
is a new action for the copytool to perform, and the FIDs to
use.

These are handled by ct_restore_v2, ct_archive_v2 and
ct_migrate_v2 which use the new llapi interface to operate.

A new line in a work file looks like:
ARCHIVE [0x200000403:0xd:0x0] [0x200000403:0xd:0x0]
MIGRATE [0x200000403:0xd:0x0] [0x200000407:0xf:0x0]
RESTORE [0x200000403:0xd:0x0] [0x200000403:0xd:0x0]

Signed-off-by: Ben Evans <jevans@cray.com>
Test-Parameters: testlist=sanity-hsm
Change-Id: I25acc7159c3c4d470edc321aba90b08e1c6b8cae
Reviewed-on: https://es-gerrit.dev.cray.com/155418
Reviewed-by: Nikitas Angelinas <nangelinas@cray.com>
Reviewed-by: Nathan Rutman <nrutman@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>