LU-11208 tests: add version check to sanity tests
sanity test 27G was added to Lustre tag 2.11.51. Thus, we
need to check that the server is 2.11.51 or later before
running test 27G (LU-11208).

sanity test 239A, change version check to 2.10.4 (LU-10230).

sanity test 311 was modified in Lustre tag 2.12.51. We need
to check that the server is 2.12.51 or later (LU-11965).

sanity test 317 was added to Lustre tag 2.11.53.
We need to check that the server is 2.11.53 or later before
running test 317 (LU-11778).

Fixes: 37f6357a5c9f ("LU-10629 lod: Clear OST pool with setstripe")
Fixes: 0ba690a526be ("LU-7251 osp: do not assign commit callback to
                      every thandle")
Fixes: a531ab5f38a6 ("LU-11605 osp: max_create_count and create_count
                      changes")
Fixes: 6115eb7fd5 ("LU-10370 ofd: truncate does not update blocks
                    count on client")

This patch is a back port from
Lustre-commit: f7155420024e142c86ec4a589ba4ba78679c9c8d
Lustre-change: https://review.whamcloud.com/33420

Test-Parameters: trivial
Signed-off-by: James Nunez <jnunez@whamcloud.com>
Change-Id: I3be45bd0bde5ec041fefef2559656ec74448dffa
Reviewed-by: Wei Liu <sarah@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Signed-off-by: Minh Diep <mdiep@whamcloud.com>
Reviewed-on: https://review.whamcloud.com/34535
Tested-by: Jenkins
Tested-by: Maloo <maloo@whamcloud.com>