LUS-7240 lnet: Cleanup lnet_get_rtr_pool_cfg
The cfs_percpt_for_each loop contains an off-by-one error that causes
memory corruption. In addition, the way these loops are nested results
in unnecessary iterations. We only need to iterate through the cpts
until we match the cpt number passed as an argument. At that point we
want to copy the router buffer pools for that cpt.

Signed-off-by: Chris Horn <hornc@cray.com>
Change-Id: I5e47aac7f45b5563683f07a35ba79e43fb482e2d
Reviewed-on: https://es-gerrit.dev.cray.com/155026
Reviewed-by: Alexander Boyko <c17825@cray.com>
Reviewed-by: Alexey Lyashkov <c17817@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>