LUS-7242 tests: fix interop with old clients
We are unable to run the tests on the setup with mixed
clients versions installed because the clients older than
2.8.54 do not have set_hostid() added by LU-7134.

The hostid is to be set on all sever nodes including the
facets falover nodes. This is not done currently, and tests
fail on the setup like the following: ost1_HOST=host1
ost1failover_HOST=host2.

Patch fixes formatall() to call set_hostid() on server nodes
only, including the failover nodes.

Fixes: 7b654fc673 ("LU-7134 utils: Ensure hostid set for ZFS")

Signed-off-by: Elena Gryaznova <c17455@cray.com>
Change-Id: I86ff18815f8e295d460b92b0e5963c3cec042105
Reviewed-on: https://es-gerrit.dev.cray.com/155032
Reviewed-by: Alexander Boyko <c17825@cray.com>
Reviewed-by: Vitaly Fertman <c17818@cray.com>