LUS-7334 kfilnd: Incorrect transaction target NID
kfilnd_recv() is called by LNet when the target of the LNetPut/Get
receives a Put/Get request. kfilnd maintains state of this current
receive message in a transaction structure. If the receive message is
a KFILND_MSG_PUT_REQ or KFILND_MSG_GET_REQ, the target of the
LNetPut/Get needs to perform a future RMA operation to progress the
LNetPut/Get. In order to perform the future RMA operation, the target
of the LNetPut/Get needs to set the transaction target NID field to
the initiator of the LNetPut/Get request. This is the initiator field
in the LNet message handed to kfilnd when kfilnd_recv() is called.

Signed-off-by: Ian Ziemba <iziemba@cray.com>
Change-Id: Ib04187bad27d9f6e13e4be0c3c3947669998e281
Reviewed-on: https://es-gerrit.dev.cray.com/155148
Reviewed-by: Frank Zago <fzago@cray.com>