LUS-7335 kfilnd: Basic NID hash implementation
When kfilnd uses a kfabric address vector (AV) to translate a node and
service string into a provider specific address, a KFI address
(kfi_addr_t) is returned to the user as an opaque handle to this
specific provider address. It is expected that the kfabric user caches
the returned KFI address for this specific node and service string.

Since kfilnd allocates the node and service string from a LNet NID,
kfilnd can associate the returned KFI address with a specific LNet
NID instead of a node and service string.

The NID hash takes a LNet NID and translates to a KFI address. If a
KFI address is not found in the NID hash, a KFI address is allocated
from the AV, and then inserted into the hash for future use.

With this patch, kfilnd can communicate across the network (i.e. not
limited to loopback communication only).

Removing entries from the NID hash and AV is still todo.

Signed-off-by: Ian Ziemba <iziemba@cray.com>
Change-Id: Ic956fc73622e53f7892bfb48d53f11cda8d2c46b
Reviewed-on: https://es-gerrit.dev.cray.com/155234
Reviewed-by: Frank Zago <fzago@cray.com>