LUS-7343 kfilnd: Update NID hash on msg receive
Each LNet KFILND device is compromised of multiple RX contexts. A RX
context is an unique, receive only endpoint that shares the same base
address as the parent, scalable endpoint (SEP). The base address of
the SEP matches the LNet KFILND device NID. A specific RX context can
be targeted by using the base address plus a specific RX context
index.

In order to engage multiple RX contexts, information about the number
of allocated RX contexts a LNet KFILND device has needs to be shared.
Instead of sharing the number of RX contexts a node has allocated to
all peers in the network, a local node tells peers which local RX
context the peer should target when initiating a KFILND transaction.
This information is stored in the kfm_prefer_rx of the KFILND header.

Signed-off-by: Ian Ziemba <iziemba@cray.com>
Change-Id: Ie63b494e49bbe82c1b5e63db7fb7de5a9df4d045
Reviewed-on: https://es-gerrit.dev.cray.com/155235
Reviewed-by: Frank Zago <fzago@cray.com>