LUS-7359 tests: customise dom-performance
dom-performance compares the performance benefits
between the normal file and DoM file.
However it is also interesting to demostrate that if
the file grows beyond the DoM size the performance
for it is not worse compared to the normal file.

Patch allows to set file size beyond the DoM size.

Test-parameters: testslist=dom-performance/IOR

Signed-off-by: Elena Gryaznova <c17455@cray.com>
Change-Id: I9d21729ebc2eef35b7d3896dac7cf10285cd88d3
Reviewed-on: https://es-gerrit.dev.cray.com/155191
Reviewed-by: Alexey Lyashkov <c17817@cray.com>
Reviewed-by: Alexander Zarochentsev <c17826@cray.com>