LUS-7362 kfilnd: Rework checksum calculation
The previous implementation did not fold the partial checksum to
generate a 16-bit checksum. This forces the addition of new fields to
the KFILND header to align the header to a 32-bit boundary.

The current implementation performs a 16-bit checksum across the
entire message. When the message is received, the target calculates a
checksum across the entire message. If the result is zero, the
checksum was validated successfully. In addition, new KFILND header
fields can be added without forcing the header to align to 32 bits.

Signed-off-by: Ian Ziemba <iziemba@cray.com>
Change-Id: I2570f2fc21f14d02d6036ac6a7d7ff008c846db4
Reviewed-on: https://es-gerrit.dev.cray.com/155228
Reviewed-by: Frank Zago <fzago@cray.com>