LUS-7357 kfilnd: Send peer RMA RX context
When KFILND uses its bulk protocol, the initiator of the LNet Put/Get
allocates a MR and binds the MR to a RX context. By binding the MR to
a RX context, the MR becomes targetable on the fabric.

In order for the target of the LNet Put/Get to target this MR, it
needs the MR key and the RX context index to successfully complete the
RMA read/write operation.

Since the preferred RX value (sent in the header) may not be the
correct RX context for the RMA read/write, pass along a RMA RX context
value to the peer to be used for a future RMA operation. The reason
why the preferred RX value may not be correct is due to the
possibility that it may have been updated at the target before the
RMA operation occurs.

Signed-off-by: Ian Ziemba <iziemba@cray.com>
Change-Id: I2336d5fe4afb821108950b20efd44d0dabec8b51
Reviewed-on: https://es-gerrit.dev.cray.com/155236
Reviewed-by: Frank Zago <fzago@cray.com>