LUS-7392 kfilnd: Remove LNet NID on TX error
Since KFI LND is connectionless, KFI LND cannot rely on connection
management to know if a peer is gone. The only indication KFI LND
receives which may point to the peer being gone is if transmit
operations start returning errors. Thus, all transmit errors are
treated as the peer being gone. This requires the LNet NID entry to be
removed from the KFI LND device NID hash. If the peer is not gone,
removing the LNet NID entry from the hash will cause the LNet NID to
be reinserted into the NID hash on the next transaction to the peer.

To test this functionality, KFI LND uses the CFS failure injection
framework. Three new values were defined.
1. 0xF100: Fail KFI send operation.
2. 0xF101: Fail KFI read operation.
3. 0xF102: Fail KFI write operation.

To inject a failure, one of the above values needs to be written to
/sys/kernel/debug/lnet/fail_loc. Note that operations will continue to
fail until the value is cleared.

Injecting a KFI send, read, or write error will cause the LND to
progress the transaction state machine to failure without issuing a
data operation.

Signed-off-by: Ian Ziemba <iziemba@cray.com>
Change-Id: I4fed11af03d09b81daf979fea10a88c2762f5556
Reviewed-on: https://es-gerrit.dev.cray.com/155243
Reviewed-by: Frank Zago <fzago@cray.com>