LUS-7435 llite: release active extent on sync write commit
Processes can wait forever in osc_extent_wait() for the extent state
to change because the extent write is not started before the wait
begins. A kernel change to generic_write_sync() modified it to check
IOCB_DSYNC instead of O_SYNC. Thus an active extent is not released
(written) in osc_io_commit_async() in the synchronous case.

This mod was originally made to 2.7 for SLES15 (LUS-1700). It should
have also been included in 2.11 but was overlooked.

Signed-off-by: Ann Koehler <amk@cray.com>
Change-Id: I3826b29769c29e4d8c3774cb97c12418bcc7b817
Reviewed-on: https://es-gerrit.dev.cray.com/155261
Reviewed-by: Ben Evans <jevans@cray.com>
Reviewed-by: Andriy Skulysh <c17819@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>