LUS-7442 kfilnd: Restrict LNet NI to assigned CPTs
System admins have the ability to restrict the number of CPTs an LNet
interface. Prior to this commit, KFI LND was not respecting the CPTs
assigned to the LNet NI.

The number of allocated KFI LND endpoints is each to the number of
CPTS the corresponding LNet NI is configured for.

When LNet calls kfilnd_send(), it is possible that the thread of
execution is running on a CPU assigned to a CPT that does not belong
to the LNet NI. If this is the case, a CPT is assigned to the
transaction based on the target's NID.

Signed-off-by: Ian Ziemba <iziemba@cray.com>
Change-Id: I3873f1b25548f650d18093e7a5a9447fbb6f0879
Reviewed-on: https://es-gerrit.dev.cray.com/155270
Reviewed-by: Frank Zago <fzago@cray.com>